% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_logit_data.R
\name{simulate_logit_data}
\alias{simulate_logit_data}
\title{Function for simulating binary data}
\usage{
simulate_logit_data(model = "normal", params = list())
}
\arguments{
\item{model}{a string which provides the distribution according to which the
feature matrix, X, will be drawn. Default value is "normal". Must be one of:
\itemize{
  \item "normal".
  \item "bernoulli".
}}

\item{params}{an optional list of model parameters that can be specified.
Supported parameters include:
\itemize{
  \item n - an integer specifying the number of observations to simulate.
  \item theta - a vector containing the true values of the logistic
        regression coefficients.
  \item intercept - a boolean for whether or not the model should include an
        intercept.
  \item sigma - a matrix to be used as the covariance matrix if the model is
        normal.
  \item px - a vector of probabilities to use for the draws if the model is
        bernoulli.
}
Default parameters for both models are n = 10000, theta = c(3, 3, 0),
and intercept = TRUE.
Default model is the normal model with the identity covariance matrix. If
bernoulli model is specified, default probabilities are 0.5 for all features.}
}
\value{
a list with the response vector, feature matrix, and logistic
  regression coefficients. Details on the elements of the returned list:
  \itemize{
    \item y - vector of binary response variables (elements are 1 or -1).
    \item x - feature matrix.
    \item theta - vector of coefficients for the logistic regression model
      logit(p(y = 1)/(1 - p(y = 1))) = x * theta.
  }
}
\description{
This function simulates binary data for a logistic regression model.
It currently supports drawing the feature matrix from a normal or bernoulli
distribution. These models are inspired by these two papers on coresets:
\itemize{
  \item \href{https://arxiv.org/abs/1605.06423}{Coresets for Bayesian
  Logistic Regression}.
  \item \href{https://arxiv.org/abs/1710.05053}{Automated Scalable Bayesian
  Inference via Hilbert Coresets}.
}
}
\examples{
## use all defaults (gives a normal model)
simulate_logit_data()

## use defaults for bernoulli model
simulate_logit_data(model = "bernoulli")

# Possible customizations for both models

## set number of independent observations to 1000
simulate_logit_data(params = list(n = 1000))

## set true logistic regression coefficients to c(1, 1, 1)
simulate_logit_data(params = list(theta = c(1, 1, 1)))

## return data without intercept column
simulate_logit_data(params = list(intercept = FALSE))

# Customizations for normal model

## change covariance matrix
simulate_logit_data(params = list(sigma = diag(10, 2)))

# Customizations for bernoulli model

## change probability vector
simulate_logit_data(params = list(px = c(0.1, 0.2, 0.3)))
}
